Remove RpmGetSrpm and just provide RpmSpecFromFile (#3)
* Remove RpmGetSrpm and just provide RpmSpecFromFile

This way we don't have to change rpmtools if we change the way the SRPM
file is retrieved (e.g. if the file is already downloaded or is
downloaded through ftp:// or other protocols).

* Add tests in rpmtools_test.go

* Do not prefix everything with Rpm

* Export common tags in their own members

Strings are easy to mistype and it is too easy to use RpmSpec wrongly
(e.g. writing "source" instead of "sources" or "buildrequires" instead
of "buildRequires", etc.). This patch extracts common tags in separate
fields of RpmSpec, so we can use them easily.